Remove hide/un-hide UI code including all related support classes
We originally needed this code to fulfill https://www.pivotaltracker.com/projects/1355218/stories/104476572
"As a user, I want to make the book title bar go away when I don't need it, so that I can focus on reading the book."

The Readium devs have since improved UI toggling:
"Change toolbar so it appears only when the mouse is over it or it gets the focus · Issue #182 · readium/readium-js-viewer"
https://github.com/readium/readium-js-viewer/issues/182

Our code clashes with theirs, and in any case we would want to use theirs for smoother operation and greater ease of maintenance.

Removed EpubReaderDoppelganger.js completely because it was only needed for UI toggling.
Removed DltsCss.js because the classes and selectors it contained were only used by EpubReaderDoppelganger.js.
Likewise removed all classes and selectors in ReadiumCss.js that were only referenced by EpubReaderDoppelganger.js.

Tagged the commit before this one: last-commit-with-our-custom-ui-toggling-code
...in case we need to bring back this doppelganger class.